Changed all endpoints in Employees Controller to run Async methods and all endpoints return status codes

+++ MusicAPI/MusicAPI/Controllers/EmployeesController.cs
@@ -22,38 +22,74 @@ namespace MusicAPI.Controllers
         }
 
         [HttpGet]
-        public IEnumerable<Personal> Get()
+        public async Task<IActionResult> Get()
         {
-            return _context.Personal.ToList();
+            //IAction Result is a way of returning status codes along with data.
+            //async needs 
+            return Ok( await _context.Personal.ToListAsync());
         }
 
         [HttpGet("{id}")]
-        public Personal Get(string id)
+        public async Task<IActionResult> Get(string id)
         {
-            return _context.Personal.Find(id);
+            var person = await _context.Personal.FindAsync(id);
+            if(person == null)
+            {
+                return NotFound("No Record Found"); 
+            }
+            else
+            {
+                return Ok(person);
+            }
         }
 
         [HttpPost]
-        public void Post([FromBody] Personal personal)
+        public async Task<IActionResult> Post([FromBody] Personal personal)
         {
-            _context.Personal.Add(personal);
-            _context.SaveChanges();
+            if(personal != null)
+            {
+                await _context.Personal.AddAsync(personal);
+                await _context.SaveChangesAsync();
+
+                return StatusCode(StatusCodes.Status201Created);
+            }
+            else
+            {
+                return NotFound("Record is not complete"); 
+            }
         }
         
         [HttpPut("{id}")]
-        public void Put(int id, [FromBody] Personal personal)
+        public async Task<IActionResult> Put(int id, [FromBody] Personal personal)
         {
-            var person = _context.Personal.Find(id);
-            person.FirstName = personal.FirstName;
-            person.LastName = personal.LastName;
-            _context.SaveChanges();
+            var person = await _context.Personal.FindAsync(id);
+            if(person == null)
+            {
+                return NotFound("No record found with that Id");
+            }
+            else
+            {
+                person.FirstName = personal.FirstName;
+                person.LastName = personal.LastName;
+                await _context.SaveChangesAsync();
+                return Ok("Record Updated Successfully");
+            }
         }
         [HttpDelete("{id}")]
-        public void Delete(int id)
+        public async Task<IActionResult> Delete(int id)
         {
-            var person = _context.Personal.Find(id);
-            _context.Personal.Remove(person);
-            _context.SaveChanges();
+            var person = await _context.Personal.FindAsync(id);
+            if(person == null)
+            {
+                return NotFound("That record was not found"); 
+            }
+            else
+            {
+                _context.Personal.Remove(person);
+                await _context.SaveChangesAsync();
+                return Ok("Record Deleted");
+            }
+            
         }
     }
 }
